+++ src/Models/Seeds/ServiceSeeder.php
@@ -46,16 +46,6 @@ class ServiceSeeder extends BaseModelSeeder
             'config'      => ['container' => 'logs']
         ],
         [
-            'name'        => 'scripts',
-            'label'       => 'Local Script Storage',
-            'description' => 'Service for accessing local script storage.',
-            'is_active'   => true,
-            'type'        => 'local_file',
-            'mutable'     => true,
-            'deletable'   => true,
-            'config'      => ['container' => 'scripts']
-        ],
-        [
             'name'        => 'db',
             'label'       => 'Local SQL Database',
             'description' => 'Service for accessing local SQLite database.',